backend/requirements.txt
fastapi
uvicorn
chromadb
pydantic
sentence-transformers
ollama
python-dotenv
redis
spacy